requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Brotli==1.0.9
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.25.0
cryptography==3.4.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-ckeditor==6.0.0
django-cloudflare-push==0.2.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-htmlmin==0.11.0
django-http2-push==0.0b2
django-js-asset==1.2.2
django-pwa==1.0.10
django-quill-editor==0.1.21
django-recaptcha==2.0.6
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.1.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
html5lib==1.1
httplib2==0.19.1
hyperframe==6.0.0
idna==2.10
msgpack==1.0.2
mysqlclient==2.0.3
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
priority==1.3.0
proto-plus==1.18.1
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-openid==2.2.5
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
wsproto==1.0.0
zstandard==0.15.2
